add PostBooksCommand and necessary comments

+++ BooksStoreClient.Common/Domain/Book.cs
@@ -1,3 +1,5 @@
 namespace Common.Domain;
 
-public sealed record Book(int Id, string Title, decimal Price, int Bookstand, int Shelf, List<Author> Authors);
+// For the price I would create Money object to ensure that no errors occur in the future
+// decimal for reading, int for saving. To simplify things here I will use int
+public sealed record Book(int Id, string Title, int Price, int Bookstand, int Shelf, List<Author> Authors);
